requirements.txt
aiohappyeyeballs==2.4.2
aiohttp==3.10.8
aiosignal==1.3.1
anyio==4.6.2
attrs==24.2.0
backoff==2.2.1
black==24.8.0
certifi==2024.8.30
charset-normalizer==3.3.2
click==8.1.7
frozenlist==1.4.1
gql==3.5.0
graphql-core==3.2.5
idna==3.10
isort==5.13.2
multidict==6.1.0
mypy-extensions==1.0.0
packaging==24.1
pathspec==0.12.1
platformdirs==4.3.6
python-dotenv==1.0.1
requests==2.32.3
sniffio==1.3.1
SQLAlchemy==2.0.35
tqdm==4.66.5
typing_extensions==4.12.2
urllib3==2.2.3
yarl==1.13.1
